% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{dircheck}
\alias{dircheck}
\title{Check that jobfolder is in working directory}
\usage{
dircheck(jobname)
}
\arguments{
\item{jobname}{Character string.}
}
\description{
Check that jobfolder is in working directory
}

