% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{get_logL}
\alias{get_logL}
\title{Get LogL score from asreml .asr file.}
\usage{
get_logL(x)
}
\arguments{
\item{x}{SNP name.}
}
\description{
Get LogL score from asreml .asr file random snp run.
}
\keyword{asreml}
\keyword{gwas}

