% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{job_setup}
\alias{job_setup}
\title{Split genotype data-fra into n_jobs pieces.}
\usage{
job_setup(snplist, n_jobs)
}
\arguments{
\item{snplist}{character vector of SNP-names.}

\item{n_jobs}{N jobs to split the job over.}
}
\description{
Split genotype data-fra into n_jobs pieces.
}

