% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{parse_results}
\alias{parse_results}
\title{Process asreml results files}
\usage{
parse_results(x, multicore = FALSE)
}
\arguments{
\item{x}{data_loop object.}
}
\description{
MAKE A NICER OUTPUT and get number of individuals per genotype class
}
\keyword{asreml}

