% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{read_genotypes}
\alias{read_genotypes}
\title{Read specified markers into the geno object.}
\usage{
read_genotypes(x, markers = NULL)
}
\arguments{
\item{x}{File name as character string.}

\item{markers}{numeric vector of columns to keep.}
}
\description{
Read specified markers into the geno object.
}

