% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{read_region}
\alias{read_region}
\title{Read spesified region of ASreml genofile.}
\usage{
read_region(variant_map, start_r, end_r, chrom_r, genofile)
}
\arguments{
\item{variant_map}{data.frame Map data.frame.}

\item{start_r}{integer. Start of region}

\item{end_r}{integer. End of reion}

\item{chrom_r}{string. Chromosome of region.}

\item{genfile}{string. Path to genofile.}
}
\value{
data.frame
}
\description{
Read spesified region of ASreml genofile.
}

