% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{readANOVAASREML}
\alias{readANOVAASREML}
\title{Process asreml ANOVA output}
\usage{
readANOVAASREML(name)
}
\arguments{
\item{name}{Dummy variable.}
}
\description{
function for process the ANOVA (Don't require changes)
}
\keyword{asreml}

