% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{readANOVAASREML_multicore}
\alias{readANOVAASREML_multicore}
\title{Process asreml ANOVA output}
\usage{
readANOVAASREML_multicore(run, SNP)
}
\arguments{
\item{run}{integer. Run number.}

\item{SNP}{Character. SNP identifier.}
}
\description{
function for process the ANOVA in multi-core setting
}
\keyword{GWAS}

