% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{split_n_run}
\alias{split_n_run}
\title{split and run gwas files on cluster.}
\usage{
split_n_run(run, runs, jobname, phenofile, pedigree, geno, dirlist)
}
\arguments{
\item{run}{integer Run number.}

\item{runs}{list Character list of SNPs to run.}

\item{jobname}{string. Name of job}

\item{phenofile}{string. Phenotype file}

\item{geno}{data.frame. Genotype data.frame}

\item{dirlist}{list. List of new and old dirs as arguments.}
}
\description{
split and run gwas files on cluster.
}
\keyword{asreml}
\keyword{gwas}

