% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results_Tim.R
\name{parse_results_Tim}
\alias{parse_results_Tim}
\title{Process asreml results files}
\usage{
parse_results_Tim(x, multicore = FALSE)
}
\arguments{
\item{x}{data_loop object.}
}
\description{
MAKE A NICER OUTPUT and get number of individuals per genotype class
}
\keyword{asreml}

