% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_utils_2.R
\name{subset_common}
\alias{subset_common}
\title{Format genotypes data frame.}
\usage{
subset_common(x, fraction = FALSE)
}
\arguments{
\item{x}{plink raw format data frame.}

\item{fraction}{Locical. Will pick a random fraction of markes for
testing if TRUE. Default [FALSE]}
}
\description{
Format genotypes data frame. Make random subset of markers if fraction = TRUE.
Find common pheno/geno animals.
}
\keyword{asreml}
\keyword{gwas}

